+++ leetcode/SingleRowKeyboard.java
@@ -9,13 +9,13 @@ class SingleRowKeyboard {
         public int calculateTime(String keyboard, String word) {
             int[] chars = new int[26];
             for (int i = 0; i < keyboard.length(); i++) {
-                int asciiRelative = (int) (keyboard.charAt(i) - 'a');
+                int asciiRelative = keyboard.charAt(i) - 'a';
                 chars[asciiRelative] = i;
             }
 
             int time = 0, prev = 0;
             for (int i = 0; i < word.length(); i++) {
-                int asciiRelative = (int) (word.charAt(i) - 'a');
+                int asciiRelative = word.charAt(i) - 'a';
                 time += Math.abs(chars[asciiRelative] - prev);
                 prev = chars[asciiRelative];
             }
